Create a data set of polygons for a set of letters from a specified fontfamily

Create a data set of polygons for a set of letters from a specified fontfamily

createPolygons(letters, font, fontsize = 400, dim = c(720, 720), scale = FALSE)

Arguments

letters

set of characters for which polygons are to be created

font

character describing the name of a font - use e.g the command `extrafont::fonts()` to check on available fonts and installing other fonts.

fontsize

size of letter to be created - larger means higher resolution, but also bigger result sets

dim

dimensions of the box in which the created letter is supposed to fit

scale

scale the values along the y axis to use result in geom_logo

Value

data set compatible to work with geom_logo

Examples

# \donttest{
# check that all letters and digits are nicely shaped:
new_alphabet <- createPolygons(c("f", "g", "W", "*"), font="Garamond")
#> Warning: no font could be found for family "Garamond"
#> Warning: no font could be found for family "Garamond"
#> Warning: no font could be found for family "Garamond"
#> Error in createPolygons(c("f", "g", "W", "*"), font = "Garamond"): Font 'Garamond' not found.

library(ggplot2)
qplot(x,y, geom="polygon", data=new_alphabet, facets=~group)
#> Warning: `qplot()` was deprecated in ggplot2 3.4.0.
#> Error in eval(expr, envir, enclos): object 'new_alphabet' not found
# }